added leads and property module.

+++ server/api/properties/[id]/index.delete.ts
@@ -1,7 +1,12 @@
-import { deleteProperties, zPropertyIdsParam } from '~~/server/utils/properties'
+import {
+  assertCanDeleteProperty,
+  deleteProperties,
+  zPropertyIdsParam
+} from '~~/server/utils/properties'
 
 export default defineEventHandler(async event => {
-  await getCurrentUser(event)
+  const user = await getCurrentUser(event)
+  assertCanDeleteProperty(user)
   const params = await validate({ id: getRouterParam(event, 'id') }, zPropertyIdsParam)
   const ids = params.id
   return deleteProperties(ids)
